Reverted change of _id to object id's. _id is now once again, and forever more a string.
This makes the code more simple and the behavior of _id's
more consistent. Previously, _id was object id and all
association keys were strings. This made a mess of things.

The alternative is to force the user to convert any id param
in rails or merb to object id's before querying and that seems
like too much of a pain.