Fixed MongoMapper's pagination proxy respond_to?
Proxy was returning the right values for its association methods, however it was undefining respond_to? at the top, so it was using method_missing for that, which gives incorrect results.

Fixed things so it doesn undefine respond_to, and it properly proxy's respond_to around itself and the subject.