Consider polymorphic belongs_to in counter_cache
When there's a polymorphic belongs_to, counter cache fails to work because doesn't exists a class named as the polymorphic association. On polymorphic associations is not possible to know previously if the parent model has the correct attribute to increment, so we need to check in the callback if the parent model responds to the method.