Refactor scoping and make default_scopes thread-safe
We've used `active_scopes` and `default_scopes` to store scopes in the current
context. However, it has been thread-unsafe because we clear `default_scopes`
to unscope.
Moreover, `default_scopes` is the master data of default scopes, and we shouldn't
clear it.

Therefore, this commit introduces `current_context` as a thread local variable to
store the current scoped context in it:

- When `unscoped { ... }` is called, `current_context` stores a plain query object
- When `with_scope(...) { ... }` is called, `current_context` stores a query object
  scoped with `default_scopes` and the method's arguments
- Otherwise, it stores nothing - nil, so we need to build a new query with
  `default_scopes`

`current_context` includes active scopes naturally, so this commit removes the method.